% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_mng.R
\name{duckdb_mng}
\alias{duckdb_mng}
\title{DuckDB Manager Class
This R6 class provides a simple interface for managing a DuckDB connection
and executing queries. It includes methods for connecting to the database,
checking the connection status, executing SQL queries, and disconnecting
from the database.}
\description{
DuckDB Manager Class
This R6 class provides a simple interface for managing a DuckDB connection
and executing queries. It includes methods for connecting to the database,
checking the connection status, executing SQL queries, and disconnecting
from the database.

DuckDB Manager Class
This R6 class provides a simple interface for managing a DuckDB connection
and executing queries. It includes methods for connecting to the database,
checking the connection status, executing SQL queries, and disconnecting
from the database.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-duckdb_mng-new}{\code{duckdb_mng$new()}}
\item \href{#method-duckdb_mng-print}{\code{duckdb_mng$print()}}
\item \href{#method-duckdb_mng-db_connect}{\code{duckdb_mng$db_connect()}}
\item \href{#method-duckdb_mng-is_connected}{\code{duckdb_mng$is_connected()}}
\item \href{#method-duckdb_mng-db_get_query}{\code{duckdb_mng$db_get_query()}}
\item \href{#method-duckdb_mng-db_disconnect}{\code{duckdb_mng$db_disconnect()}}
\item \href{#method-duckdb_mng-clone}{\code{duckdb_mng$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_mng-new"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_mng-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_mng$new(db_path = ":memory:")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_mng-print"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_mng-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_mng$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_mng-db_connect"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_mng-db_connect}{}}}
\subsection{Method \code{db_connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_mng$db_connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_mng-is_connected"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_mng-is_connected}{}}}
\subsection{Method \code{is_connected()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_mng$is_connected()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_mng-db_get_query"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_mng-db_get_query}{}}}
\subsection{Method \code{db_get_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_mng$db_get_query(sql_query, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_mng-db_disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_mng-db_disconnect}{}}}
\subsection{Method \code{db_disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_mng$db_disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_mng-clone"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_mng-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_mng$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
