% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{get_brand_theme}
\alias{get_brand_theme}
\title{Get the application theme}
\usage{
get_brand_theme()
}
\value{
a bslib theme object
}
\description{
A function to get the application theme from the brand yaml file.
}
