% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_funcs.R
\name{get_db_store}
\alias{get_db_store}
\title{Get Database Manager
A helper function to create and return an instance of the DuckDB manager class.}
\usage{
get_db_store(
  db_path = system.file("extdata", "mtcars.duckdb", package = "golem_shiny")
)
}
\arguments{
\item{db_path}{The file path to the DuckDB database. Defaults to the bundled
mtcars.duckdb file in the package's extdata directory.}
}
\value{
An instance of the duckdb_mng class, connected to the specified database
}
\description{
Get Database Manager
A helper function to create and return an instance of the DuckDB manager class.
}
