% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_table_server}
\alias{mod_table_server}
\title{reports Server Function}
\usage{
mod_table_server(id)
}
\arguments{
\item{id:}{a random chr string but similar to ui module id}
}
\description{
A shiny Module.
}
