% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_funcs.R
\name{get_db_store}
\alias{get_db_store}
\title{Get Database Manager
A helper function to create and return an instance of the DuckDB manager class.}
\usage{
get_db_store(
  package = "golem_shiny",
  folder = "extdata",
  filename = "mtcars.duckdb"
)
}
\arguments{
\item{package}{The package name containing the DuckDB database file.}

\item{folder}{The folder within the package where the DuckDB database file is located.}

\item{filename}{The name of the DuckDB database file.}
}
\value{
An instance of the duckdb_mng class, connected to the specified database
}
\description{
Get Database Manager
A helper function to create and return an instance of the DuckDB manager class.
}
